fix(linker): we introducing garbage value in zig with this one :fire::fire::fire:
the most facisnating bug i've seen in quite some time.
my dumbass thought that the entire .bss section is dedicated to stack
but it turns out that it also contains other uninitialized data, so i
stored the sp as the .bss end which was overwriting the terminal
writer's row and column variable which broke the writing functionality.

when calculating any thing the zig compiler automatically inserts
overflow checks, the value that was overwritten by the stack was causing
that overflow. when hit with the overflow, panic gets called, but the
panic itself contains call to terminal writer. idk how but then zig does
sm shit and runs an invalid opcode that causes the kernel to crash.